requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
autoflake==1.4
Automat==20.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
Django==3.2.2
django-redis-cache==3.0.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
docformatter==1.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.1.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
openapi-codec==1.3.2
pep8==1.7.1
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyformat==0.7
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
unify==0.5
untokenize==0.1.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.4.0
